requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
playsound>=1.3.0